+++ .env.ethusd
@@ -1,0 +1,12 @@
+BLOCK_MATURITY=0.5
+BYPASS_KRAKEN_API=false
+GOAL=75000
+GOAL_START=10607
+KRAKEN_API_KEY=emuS1t6BeMeBE91mNsnsYqaVbHlY5imEmnmUrY/DzT6Rg3CEqEYlxwad
+KRAKEN_API_SECRET=0kw9rMld9sVcCLlwqfTIz4NPS+9suy5r7zrPTNRD+s92WGm6BL1dk7jdSUBC/oy+T8fU73B9ZY8NKvE+kgFEVQ==
+MAX_BET=4600
+PAIR=ETHUSD
+RESERVE=6000
+SLACK_BOT_TOKEN=xoxb-1462120298480-1772780122162-RCMM7ilNG7Cu7X2Vb31d8rCV
+SLACK_CHANNEL=C01NZ0WERCL
+TARGET_PROFIT=0.01
