+++ src/app.ts
@@ -2,14 +2,14 @@ import { AssetWatcher } from './assetWatcher/assetWatcher'
 import { Bot } from './bot'
 import { BuyAnalyst } from './analysts/buyAnalyst'
 import { BotConfig, config } from './common/config'
-import { FullProfitBot } from './fullProfitBot'
+import { FullProfitBot } from './bots/fullProfitBot'
 import { formatMoney, formatNumber, positionId } from './common/utils'
 import { KrakenService } from './kraken/krakenService'
 import { logger } from './common/logger'
 import { PositionsService } from './positions/positions.service'
 import { round } from 'lodash'
 import { SellAnalyst } from './analysts/sellAnalyst'
-import { TrailingStopLossBot } from './trailingStopLossBot'
+import { TrailingStopLossBot } from './bots/trailingStopLossBot'
 import connect from './common/db/connect'
 import KrakenClient from 'kraken-api'
