+++ src/bots/utils.ts
@@ -1,0 +1,38 @@
+import { round } from 'lodash'
+import { Position } from '../positions/position.interface'
+import { logger } from '../common/logger'
+
+export const inWinZoneByAmount = (currentBidPrice: number, targetProfit: number, volume: number, price: number, tax: number): boolean => {
+  const costs = price * volume
+  const fee = costs * tax * 2
+  const totalCosts = fee + costs
+
+  const volumeToSell = round((totalCosts / currentBidPrice), 0)
+  const expectedProfit = volume - volumeToSell
+
+  logger.debug(`Expected profit: ${expectedProfit}`)
+
+  return expectedProfit > 0 && expectedProfit >= targetProfit
+}
+
+export const inWinZoneByPercentage = (currentBidPrice: number, targetProfit: number, volume: number, price: number, tax: number): boolean => {
+  const costs = price * volume * (1 + tax)
+  const profit = currentBidPrice * volume * (1 - tax)
+  const percentage = 100 * profit / costs
+
+  logger.debug(`FullProfitBot: ${costs} -- ${profit} -- ${percentage}`)
+  return percentage > (100 + targetProfit)
+}
+
+export const inWinZone = (position: Position, currentBidPrice: number, targetProfit: number, tax: number): boolean => {
+  if(!position.buy.price || !position.buy.volume) {
+    throw new Error(`Not enough data to estimate win zone`)
+  }
+
+  if(targetProfit > 1) {
+    return inWinZoneByAmount(currentBidPrice, targetProfit, position.buy.volume, position.buy.price, tax)
+  }
+  else {
+    return inWinZoneByPercentage(currentBidPrice, targetProfit, position.buy.volume, position.buy.price, tax)
+  }
+}
