+++ src/test.ts
@@ -10,7 +10,7 @@ import moment from 'moment'
 import { config } from './common/config'
 import { ProfitsRepo } from './profit/profit.repo'
 import { KrakenService } from './kraken/krakenService'
-import { TrailingStopLossBot } from './trailingStopLossBot'
+import { TrailingStopLossBot } from './bots/trailingStopLossBot'
 import { Bot } from './bot'
 import { flatMap } from 'lodash'
 import { AssetWatcher } from './assetWatcher/assetWatcher'
